% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{Matrix Folds/species}
\format{
A data frame with 1072 rows and 210 variables
}
\source{
<https://doi.org/10.1111/evo.14550>
}
\usage{
my_dataset
}
\description{
Jeu de donnees biologique.
}
\keyword{datasets}
